replace LineEndMarker by End() Method
